modify

+++ ethNode/app/ethClient.py
@@ -50,7 +50,7 @@ class Client(object):
             gasprice=tx_dict.get("gasPrice"),
             startgas=tx_dict.get("gas"),
             to=tx_dict.get("to"),
-            value=tx_dict.get("value")*(10**8),
+            value=int(tx_dict.get("value")*(10**8)),
             data=binascii.unhexlify(tx_dict.get("data")[2:]))
 
         UnsignedTransaction = Transaction.exclude(['v', 'r', 's'])
